Reorganize paper to remove "existing mechanisations" section.
The idea is to separate the content of the section such that the
motivational parts for the benchmark is merged into the introduction and
the parts that motivate why we chose these challenges are merged into the
section about the design of the benchmark.

For now, I have just moved everything into the introduction to clean
up later.