Reorganise the short paper.
This undoes the idea of splitting the related work in two sections,
since this never really worked out well.
This also cleans up the introduction by moving things elsewhere.

The sections about each challenge are not integrated yet, and neither
are the sections on the overall design and possible extensions.